requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-CORS==3.0.10
python-dotenv==0.21.1
SQLAlchemy==1.4.47  
